Lisättiin yksi ampumistesti sekä muokattiin luokkia
Muutettiin lentoradan iteroiRata() metodia
niin, ettei se palauta mitään ja operoi vain
luokan jäsen Ammukseen. Muutettiin myös testit
huomioimaan tämä muutos.
Lisättiin yksi testi Tykkipelin pelaajaAmpuu()
metodiin, jossa käytetään kaavaa vinon heitto-
liikkeen kantaman laskuun 
[http://fi.wikipedia.org/wiki/Heittoliike].

+++ netbeans/tykkipeli/src/tykkipeli/controller/Tykkipeli.java
@@ -73,7 +73,7 @@ public class Tykkipeli {
 	}
     }
     
-    public Lentorata pelaajaAmpuu(Pelaaja pelaaja, double koro, int teho) {
+    public Lentorata pelaajaAmpuu(Pelaaja pelaaja, double koro, double teho) {
 	pelaaja.setKoro(koro);
 	pelaaja.setTeho(teho);
 	Point alkupsijainti = pelaaja.getTykki().getSijainti();
